# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk Sistem Manajemen Work Order.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dashboard/stats": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard Statistics",
                "operationId": "654e1617f74d2b9e0654ec77620d960b",
                "parameters": [
                    {
                        "name": "assigned_operator_id",
                        "in": "query",
                        "description": "ID Operator",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totalWorkOrders": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "pending": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "inProgress": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "completed": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "canceled": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}